Improve build_cts_json.yaml workflow security
Limits the increased permissions to a dedicated job which runs only on push events.

The previous implementation was already enough to protect against malicious PRs from forks.
However, if Dependabot was used on the repository, it would have increased its default
read-only permissions to write permissions, making it easier for compromised dependencies
to directly affect the repository.